chrom : move CachedScope from cache.rs to scope.rs
